+++ tensor_graphs_cpp/kernels/opencl/gelu/gelu.cl
@@ -1,7 +1,7 @@
 // tensor_graphs_cpp/kernels/opencl/gelu/gelu.cl
 
 __kernel void gelu_f32_nd(__global const float* A, __global float* Out, ulong n) {
-    size_t idx = get_global_id(0);
+    ulong idx = get_global_id(0);
     if (idx < n) {
         float x = A[idx];
         // exact erf mapping: 0.5 * x * (1 + erf(x / sqrt(2)))
